Move RenderValueColumnCollText to BocValueColumnRendererBase
